requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
paystackapi>=1.0.0
gunicorn>=21.2.0

#optional
python-decouple>=3.0

psycopg2-binary>=2.9.0
